requirements.txt
python-docx
python-docx-template
streamlit
tabulate
pandas
openpyxl
Pillow
